% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dou2019_1}
\alias{dou2019_1}
\alias{dou2019_1_protein}
\title{FACS + nanoPOTS + TMT multiplexing: HeLa digests (Dou et al. 2019)}
\format{\code{dou2019_1} contains 1 set of data:
\itemize{
  \item \code{dou2019_1_protein}: an MSnSet with protein expression levels
  for 1641 proteins x 20 "cells".
}
See Details for information about data collection.}
\source{
The original data can be downloaded from the 
\href{https://pubs.acs.org/doi/10.1021/acs.analchem.9b03349}{ACS Publications}
website (Supplementary information section).
}
\usage{
dou2019_1
data("dou2019_1_protein")
}
\description{
Single-cell proteomics using nanoPOTS combined with TMT isobaric labeling. 
The data set is published as the supplementary data set 1 by Dou et al. in 
2019 (see references). The expression matrix contains 1641 proteins x 20 
single-cell HeLa digests. The samples are not truly single cells but are 
commercial Hela digest diluted to single cell amounts (0.2ng). The boost 
wells contain the same digest but at hihgher dose (10 ng).
}
\details{
The data set was downloaded from the supplementary information section of the
publisher's website (see sources). 

\strong{Protein expression data: \code{dou2019_3_protein}}

The spreadsheet is called \code{ac9b03349_si_003.xlsx} and contains 7 sheets 
from which we only took the sheet 6 (named \code{"5 - Run 1 and 2 raw data"}) 
containing the combined data of two MS runs. It corresponds to the assembly 
of MSGF+ identifications and MASIC reporter assemblies. The data is then isotope 
corrected and sum rolled-up to the protein level. Contaminant and reverse hit 
are removed from this table. This is all performed by the authors. We 
converted the data to an \code{\link{MSnSet}} object.
}
\references{
Dou, Maowei, Geremy Clair, Chia-Feng Tsai, Kerui Xu, William B. 
Chrisler, Ryan L. Sontag, Rui Zhao, et al. 2019. “High-Throughput Single Cell 
Proteomics Enabled by Multiplex Isobaric Labeling in a Nanodroplet Sample 
Preparation Platform.” Analytical Chemistry, September 
(\href{https://doi.org/10.1021/acs.analchem.9b03349}{DOI}).
}
\keyword{datasets}
