% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dou2019_3}
\alias{dou2019_3}
\alias{dou2019_3_protein}
\title{FACS + nanoPOTS + TMT multiplexing: profiling of murine cell populations
(Dou et al. 2019)}
\format{\code{dou2019_3} contains 1 set of data:
\itemize{
  \item \code{dou2019_3_protein}: an MSnSet with protein expression levels
  for 2331 proteins x 132 cells.
}
See Details for information about data collection.}
\source{
The original data can be downloaded from the 
\href{https://pubs.acs.org/doi/10.1021/acs.analchem.9b03349}{ACS Publications}
website (Supplementary information section).
}
\usage{
dou2019_3
data("dou2019_3_protein")
}
\description{
Single-cell proteomics using nanoPOTS combined with TMT isobaric labeling. 
The data set is published as the supplementary data set 3 by Dou et al. in 
2019 (see references). The expression matrix contains 2331 proteins x 132 
wells. Among the 132 wells, 72 contained single cells, corresponding to 24 
C10 cells, 24 RAW cells, and 24 SVEC. The other wells are eithers boosting 
channels (12), empty channels (36) or reference channels (12). The 
different cell types where evenly distributed across 4 nanoPOTS chips. 
Samples were 11-plexed with TMT ions.
}
\details{
The data set was downloaded from the supplementary information section of the
publisher's website (see sources). 

\strong{Protein expression data: \code{dou2019_3_protein}}

The spreadsheet is called \code{ac9b03349_si_005.xlsx} and contains 7 sheets 
from which we took only the 2nd (named \code{"01 - Raw sc protein data"}). It 
corresponds to the assembly of MSGF+ identifications and MASIC reporter 
assemblies. The data is then isotope corrected and sum rolled-up to the 
protein level. Contaminant and reverse hit are removed from this table. This 
is performed by the authors. We converted the data to an \code{\link{MSnSet}} 
object.
}
\references{
Dou, Maowei, Geremy Clair, Chia-Feng Tsai, Kerui Xu, William B. 
Chrisler, Ryan L. Sontag, Rui Zhao, et al. 2019. “High-Throughput Single Cell 
Proteomics Enabled by Multiplex Isobaric Labeling in a Nanodroplet Sample 
Preparation Platform.” Analytical Chemistry, September 
(\href{https://doi.org/10.1021/acs.analchem.9b03349}{DOI}).
}
\seealso{
\code{\link{dou2019_1}}, \code{\link{dou2019_2}}
}
\keyword{datasets}
