% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dou2019_hela_protein}
\alias{dou2019_hela_protein}
\alias{dou2019_hela_peptide}
\title{FACS + nanoPOTS + TMT: HeLa digests (Dou et al. 2019)}
\format{\itemize{
  \item \code{dou2019_hela_protein}: a \code{\link{SingleCellExperiment}}
  object with protein expression levels for 24,797 peptides x 20 "cells".
  \item \code{dou2019_hela_protein}: a \code{\link{SingleCellExperiment}}
  object with protein expression levels for 1,641 proteins x 20 "cells".
  
}
See Details for information about data collection.}
\source{
The peptide and raw data can be downloaded from the massIVE repository 
\href{ftp://massive.ucsd.edu/MSV000084110/}{MSV000084110}.

The protein data can be downloaded from the 
\href{https://pubs.acs.org/doi/10.1021/acs.analchem.9b03349}{ACS Publications}
website (Supplementary information section).
}
\usage{
data("dou2019_hela_protein")
data("dou2019_hela_peptide")
}
\description{
Single-cell proteomics using nanoPOTS combined with TMT isobaric labeling. 
The samples are not truly single cells but are commercial Hela digest diluted 
to single cell amounts (0.2ng). The boosting wells contain the same digest 
but at higher dose (10 ng).
}
\details{
\strong{Peptide expression data: \code{dou2019_hela_peptide}}

The peptide data was constructed as follows. For every MS run:
\itemize{
  \item Load the MSGF+ identification files 
  (\code{"Hela_run_*_msgfplus.mzid"}) and remove decoy and contaminant PSMs
  \item Load the MASIC quantification files 
  (\code{"Hela_run_*_ReporterIons.txt"})
  \item Combine the identification and the quantification data by matching 
  the \code{"scan.number.s"} and \code{"scanNumber"} fields, respectively
  \item Format the data as an \code{\link{MSnSet}} object keeping 
  \item Perform TMT-10 istope correction 
  \item Sum-roll the PSM intensities to peptide intensities
  \item Keep only useful feature data (peptide and protein sequence, start 
  and end position of peptide location, protein length, protein description)
  \item Extract the sample information (TMT ion, run number and experiment 
  label) from the files names. Furthermore, sample type (lysate, carrier or
  blank) and sample amount (in ng) were added from the article (see 
  Reference).
}
Next, combine all runs in a single MSnSet (matching on peptide name), replace
NA's by 0's and convert to an \code{\link{SingleCellExperiment}} object.

\strong{Protein expression data: \code{dou2019_hela_protein}}

The data set is published as the supplementary data set 1 by Dou et al. in 
2019 (see references). The expression matrix contains 1641 proteins x 20 
HeLa digests. The spreadsheet is called \code{ac9b03349_si_003.xlsx} and 
contains 7 sheets from which we only took the sheet 6 (named \code{"5 - Run 
1 and 2 raw data"}) containing the combined data of two MS runs. It 
corresponds to the assembly of MSGF+ identifications and MASIC reporter 
assemblies. The data is then isotope corrected and sum rolled-up to the 
protein level. Contaminant and reverse hit are removed from this table. This 
is all performed by the authors. We converted the data to a 
\code{\link{SingleCellExperiment}} object.
}
\references{
Dou, Maowei, Geremy Clair, Chia-Feng Tsai, Kerui Xu, William B. 
Chrisler, Ryan L. Sontag, Rui Zhao, et al. 2019. “High-Throughput Single Cell 
Proteomics Enabled by Multiplex Isobaric Labeling in a Nanodroplet Sample 
Preparation Platform.” Analytical Chemistry, September 
(\href{https://doi.org/10.1021/acs.analchem.9b03349}{DOI}).
}
\seealso{
\code{\link{dou2019_mouse_protein}}, \code{\link{dou2019_boosting_protein}}
}
\keyword{datasets}
