% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dou2019_mouse_protein}
\alias{dou2019_mouse_protein}
\alias{dou2019_mouse_peptide}
\title{FACS + nanoPOTS + TMT: profiling of murine cell populations (Dou et al. 2019)}
\format{\itemize{
  \item \code{dou2019_mouse_peptide}: a \code{\link{SingleCellExperiment}}
  with protein expression levels for 74,374 proteins x 132 samples.
  \item \code{dou2019_mouse_protein}: a \code{\link{SingleCellExperiment}}
  with protein expression levels for 2331 proteins x 132 samples.
}
See Details for information about data collection.}
\source{
The peptide and raw data can be downloaded from the massIVE repository 
\href{ftp://massive.ucsd.edu/MSV000084110/}{MSV000084110}.

The protein data can be downloaded from the 
\href{https://pubs.acs.org/doi/10.1021/acs.analchem.9b03349}{ACS Publications}
website (Supplementary information section).
}
\usage{
data("dou2019_mouse_peptide")
data("dou2019_mouse_protein")
}
\description{
Single-cell proteomics using nanoPOTS combined with TMT isobaric labeling. 
The cell types are either "Raw" (macrophage cells), "C10" (epihelial cells), 
or "SVEC" (endothelial cells). Out of the 132 wells, 72 contain single cells, 
corresponding to 24 C10 cells, 24 RAW cells, and 24 SVEC. The other wells are 
either boosting channels (12), empty channels (36) or reference channels 
(12). Boosting and reference channels are balanced (1:1:1) mixes of C10, 
SVEC, and RAW samples at 5 ng and 0.2 ng, respectively. The different cell 
types where evenly distributed across 4 nanoPOTS chips. Samples were 
11-plexed with TMT labeling.
}
\details{
\strong{Peptide expression data: \code{dou2019_mouse_peptide}}

The peptide data was constructed as follows. For every MS run:
\itemize{
  \item Load the MSGF+ identification files 
  (\code{"Single_Cell_Chip_*_msgfplus.mzid"}) and remove decoy and 
  contaminant PSMs
  \item Load the MASIC quantification files 
  (\code{"Single_Cell_Chip_*_ReporterIons.txt"})
  \item Combine the identification and the quantification data by matching 
  the \code{"scan.number.s"} and \code{"scanNumber"} fields, respectively
  \item Format the data as an \code{\link{MSnSet}} object keeping 
  \item Perform TMT-11 istope correction 
  \item Sum-roll the PSM intensities to peptide intensities
  \item Keep only useful feature data (peptide and protein sequence, start 
  and end position of peptide location, protein length, protein description)
  \item Extract the sample information (TMT ion, chip number and experiment 
  label) from the files names. Furthermore, sample type was obtained from 
  Table S1 (see reference).
}
Next, combine all runs in a single MSnSet (matching on peptide name), replace
NA's by 0's and convert to an \code{\link{SingleCellExperiment}} object.

\strong{Protein expression data: \code{dou2019_mouse_protein}}

The data set is published as the supplementary data set 3 by Dou et al. in 
2019 (see references). The expression matrix contains 2331 proteins x 132 
wells. The spreadsheet is called \code{ac9b03349_si_005.xlsx} and contains 7 
sheets from which we took only the 2nd (named \code{"01 - Raw sc protein 
data"}). It corresponds to the assembly of MSGF+ identifications and MASIC 
reporter assemblies. The data is then isotope corrected and sum rolled-up to 
the protein level. Contaminant and reverse hit are removed from this table. 
This is performed by the authors. We converted the data to a
\code{\link{SingleCellExperiment}} object.
}
\references{
Dou, Maowei, Geremy Clair, Chia-Feng Tsai, Kerui Xu, William B. 
Chrisler, Ryan L. Sontag, Rui Zhao, et al. 2019. “High-Throughput Single Cell 
Proteomics Enabled by Multiplex Isobaric Labeling in a Nanodroplet Sample 
Preparation Platform.” Analytical Chemistry, September 
(\href{https://doi.org/10.1021/acs.analchem.9b03349}{DOI}).
}
\seealso{
\code{\link{dou2019_hela_protein}}, \code{\link{dou2019_boosting_protein}}
}
\keyword{datasets}
