% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{imputeKNN}
\alias{imputeKNN}
\title{Impute missing values using K-nearest neighbours
Internal function}
\usage{
imputeKNN(obj, k = 3)
}
\description{
Impute missing values using K-nearest neighbours
Internal function
}
