% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpdata.R
\name{leduc2022}
\alias{leduc2022}
\title{Deprecated leduc2022 dataset}
\usage{
leduc2022()
}
\value{
The \code{leduc2022_pSCoPE} dataset.
}
\description{
The \code{leduc2022} dataset has been updated to include plexDIA and
pSCoPE data. The new datasets names are \code{leduc2022_pSCoPE}
(previously \code{leduc2022}) and \code{leduc2022_plexDIA} (new). See the
respective documentation pages for more information.
}
