% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{specht2018_peptide}
\alias{specht2018_peptide}
\title{SCoPE-MS + mPOP lysis upgrade: Master Mix 20180824 (Specht et al. 2018)}
\format{\itemize{
  \item \code{specht2018_peptide}: a \code{\link{SingleCellExperiment}} 
  with peptide expression levels for 1838 peptides x 220 cells.
}
See Details for information about data collection.}
\source{
The peptide data can be downloaded from the 
\href{http://slavovlab.net/mPOP/index.html}{Slavov Lab} website.
}
\usage{
data("specht2018_peptide")
}
\description{
Single cell proteomics data produced and published by Specht et al. from the 
Slavov Lab (see references). Channels are either carrier (50 cell equivalent 
of Jurkat or U-937 digestion product), empty, or single cell equivalent of 
digestion product (Jurkat or U-937).
}
\details{
\strong{Peptide expression data: \code{specht2018_peptide}}

It contains quantitative information for 1824 peptides x 220 channels. The 
peptide expression data was parsed from the \code{evidence.txt} file from
the MaxQuant output. We processed the file as follows: 
\itemize{
  \item Contaminants and reverse hits are removed
  \item Peptides with PIF (parent ion fraction) smaller than 0.8 and PEP 
  (posterior error probability) greater then 0.02 are removed
  \item Experiment sets with less than 300 identified peptides are discarded. 
  Note that all experiment sets contained more than 300 identified peptides.
  \item Some peptides were identified twice within the same experiment set 
  because of different charge states. The peptide identification with lowest 
  PEP is kept and the other(s) discarded. 
  \item Intensity data is formated to a feature (peptide) x sample 
  (combination of experiment run and TMT channel) matrix. 
  \item Peptide information is gathered in a feature data frame. PEP, PIF, 
  Score, and retention time are aggregated by taking the median. When the 
  mass is differing among the same peptide sequence, only the mass for the 
  unmodified peptide is kept.
  \item Sample information is gathered in a phenotype data frame.
}
 We finally formated the data to a \code{\link{SingleCellExperiment}} object.

\strong{Protein expression data: \code{specht2018_peptide}}

TODO
}
\references{
Specht, Harrison, Guillaume Harmange, David H. Perlman, Edward Emmott, 
Zachary Niziolek, Bogdan Budnik, and Nikolai Slavov. 2018. “Automated Sample 
Preparation for High-Throughput Single-Cell Proteomics.” bioRxiv 
(\href{https://doi.org/10.1101/399774}{DOI}).
}
\keyword{datasets}
