% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{specht2019v1_peptide}
\alias{specht2019v1_peptide}
\alias{specht2019v1_protein}
\alias{specht2019v2_peptide}
\title{FACS + SCoPE2: macrophages vs monocytes (Specht et al. 2019)}
\format{\itemize{
  \item \code{specht2019v1_peptide}: a \code{\link{SingleCellExperiment}} 
  with peptide expression levels for 7515 peptides x 682 samples
  \item \code{specht2019v1_protein}: a \code{\link{SingleCellExperiment}} 
  with protein expression levels for 2316 proteins x 356 samples (=cells)
}
See Details for information about data collection.}
\source{
The data can be downloaded from the 
\href{https://scope2.slavovlab.net/docs/data}{Slavov Lab} website. The raw 
data and quantification data can also be found in the massIVE repository 
\href{ftp://massive.ucsd.edu/MSV000083945}{MSV000083945}
}
\usage{
data("specht2019v1_protein")
data("specht2019v1_peptide")
}
\description{
Single cell proteomics data produced and published by Specht et al. from the 
Slavov Lab (see references). It contains quantitative information 356 cells. 
Cells can be either macrophages (n = 259) or monocytes (n = 97).
}
\details{
Peptide data were collected from the massIVE database (accession ID: 
MSV000083945), annotations were received during a personnal discussion but 
are also contained in version 2 of the data (see 
\code{\link{specht2019v2_peptide}}): 
\itemize{
  \item \code{ev_updated.txt}: the MaxQuant/DART-ID output file
  \item \code{annotation_fp60-97.csv}: sample phenotype annotation
  \item \code{batch_fp60-97.csv}: batch annotation
}
 
Protein data were collected from the author's website (see sources).
\itemize{
  \item \code{Proteins-processed.csv}: Proteins x single cells at 1% FDR, 
  imputed and batch corrected.
  \item \code{Cells.csv}: Annotation x single cells. Each column corresponds 
  to a single cell and the rows include relevant metadata, such as, cell type 
  if known, measurements from the isolation of the cell, and derivative 
  quantities, i.e., rRI, CVs, reliability.
}
 
\strong{Peptide expression data: \code{specht2019v1_peptide}}
 
We processed the MaxQuant output \code{ev_updated.txt} as follows:
\itemize{
  \item Keep only column of interest, that is the peptide and protein 
  information, the quantification information (TMT intensities), the 
  identification information (PEP, q-values, PIF,...) and the contamination
  information
  \item Keep only the single cell runs (experiments FP94 and FP97)
  \item Remove out reverse hits and contaminants (identified by MaxQuant), 
  and contaminated spectra (\code{PIF > 0.8})
  \item Remove peptides with low identification score (\code{FDR >= 0.01} or 
  \code{PEP >= 0.02})
  \item Remove cells with less than 300 peptides
  \item Some peptides were identified twice within the same experiment set 
  because of different charge states. The peptide identification with lowest 
  PEP is kept and the other(s) discarded.
  \item Intensity data is formated to a feature (peptide) x sample 
  (combination of experiment run and TMT channel) matrix.
  \item Zero intensities are replaced by NA's
  \item Peptide information is gathered in a feature data frame, only 
  peptide information common to all runs is kept (eg sequence, mass, 
  protein,...).
  \item Sample information is gathered in a phenotype data frame. The sample 
  information is extracted from the \code{annotation_fp60-97.csv} and 
  \code{batch_fp60-97.csv} files.
}

We finally formated the data to an \code{\link{SingleCellExperiment}} 
object. The samples in this data set are either a carrier well 
(\code{"carrier_mix"}), a normalization well (\code{"norm"}), an unused
well (\code{"unused"}), an empty well (\code{"sc_0"}), a macrophage cell
(\code{"sc_m0"}), or a monocyte cell (\code{"sc_u"}).

\strong{Protein expression data: \code{specht2019v1_protein}}
 
On top of the steps described above, the peptide expression data were further
processed by the authors to the protein data through the following steps: 
\itemize{
  \item Remove peptides that are more than 10\% the intensity of the carrier
  \item Divide peptide intensities in every channel by the reference channel
  \item Zero or infinite intensities are replaced by NA's
  \item Remove cells having median CV > 0.43, having 30th quantile of the 
  log10 transformed relative RIs smaller than -2.5, or having median of the 
  log10 transformed relative RI larger than -1.3
  \item Divide column (cells) with median intensity and divide rows with mean 
  intensity
  \item Remove rows (peptides) then columns (cells) that contain more than 
  99\% of missing data
  \item Log2 transform the data
  \item Aggregate the peptide to their corresponding protein using their 
  median value.
  \item Normalize rows by subtracting the row average and normalize columns 
  by subtracting the column median
  \item Impute missing values using K-nearest neighbors (k = 3)
  \item Correct for batch effect using the \code{\link{ComBat}} algorithm.
}

This leads to the \code{Proteins-processed.csv} file. The protein 
expression data are combined with the meta data (\code{Cells.csv}) into an 
\code{\link{SingleCellExperiment}} object. The data set contains 
macrophages (\code{"sc_m0"}) and monocytes (\code{"sc_u"}).
}
\references{
Specht, Harrison, Edward Emmott, Toni Koller, and Nikolai Slavov. 
2019. “High-Throughput Single-Cell Proteomics Quantifies the Emergence of 
Macrophage Heterogeneity.” bioRxiv (\href{https://doi.org/10.1101/665307}{DOI}).
}
\seealso{
\code{\link{specht2019v2_peptide}}, 
\code{\link{specht2019v2_protein}}
}
\keyword{datasets}
