% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhu2018MCP_peptide}
\alias{zhu2018MCP_peptide}
\title{laser dissection + nanoPOTS:  (Zhu et al. 2018, Mol Cell Proteomics)}
\format{\itemize{
  \item \code{zhu2018MCP_peptide}: a \code{\link{SingleCellExperiment}} 
  with peptide expression levels for 10074 peptides x 24 samples
}
See Details for information about data collection.}
\source{
The raw data, the idenfication and quantification results can be found in the
PRIDE repository 
\href{ftp://ftp.pride.ebi.ac.uk/pride/data/archive/2018/07/PXD008844}{PXD008844}
}
\usage{
data("zhu2018MCP_peptide")
}
\description{
Near single-cell proteomics data produced and published by Zhu et al. 
(see references). It contains quantitative information for 24 mouse brain 
sections. The samples were brains from rat pups (day 17). The slices are 12
µm thick squares of either 50, 100, or 200 µm width. 5 samples were dissected 
from the corpus callum (\code{"CC"}), 4 samples were dissected from the 
corpus collosum (\code{"CP"}), 13 samples were extracted from the cerebral 
cortex (\code{"CTX"}), and 2 samples are labeled as (\code{"Mix"}).
}
\details{
Peptide data were collected from the PRIDE repository (accession ID: 
PXD008844), sample annotations were infered from column names.

\strong{Peptide expression data: \code{zhu2018MCP_peptide}}

We processed the MaxQuant output \code{MaxQuant_Peptides.txt} as follows:
\itemize{
  \item Extract the phenotype data (sample type) from the intensity column 
  names.
  \item Remove contaminants and reverse hits 
  \item Remove peptides with a posterior error probability < 0.01
  \item Replace 0's by NA's
  \item Remove rows (peptides) that contain only NA's
}
We finally formated the data to an \code{\link{SingleCellExperiment}} 
object. The samples in this data set are coming from different brain regions
(CTX, CP, or CC) and dissected as 50, 100 or 200 µm wide squares.
}
\references{
Zhu, Ying, Maowei Dou, Paul D. Piehowski, Yiran Liang, Fangjun Wang, Rosalie 
K. Chu, William B. Chrisler, et al. 2018. “Spatially Resolved Proteome 
Mapping of Laser Capture Microdissected Tissue with Automated Sample
Transfer to Nanodroplets.” Molecular & Cellular Proteomics: MCP 17 (9): 
1864–74. \href{http://dx.doi.org/10.1074/mcp.TIR118.000686}{DOI}
}
\keyword{datasets}
