% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhu2018NC_hela_peptide}
\alias{zhu2018NC_hela_peptide}
\title{nanoPOTS: HeLa dilutions (Zhu et al. 2018, Nat Comm)}
\format{\itemize{
  \item \code{zhu2018NC_hela_peptide}: a \code{\link{SingleCellExperiment}} 
  with peptide expression levels for 30345 peptides x 21 samples
}
See Details for information about data collection.}
\source{
The raw data, the idenfication and quantification results can be found in the
PRIDE repository 
\href{ftp://ftp.pride.ebi.ac.uk/pride/data/archive/2018/01/PXD006847}{PXD006847}
}
\usage{
data("zhu2018NC_hela_peptide")
}
\description{
Near single-cell proteomics data produced and published by Zhu et al. 
(see references). It contains quantitative information for 21 samples, 
some being triplicates of HeLa dilution (~10, ~40 and ~140 cells). Remaining
samples are blanks (0 cells), lysate or cancer cell lines (THP-1 or MCF-7).
}
\details{
Peptide data were collected from the PRIDE repository (accession ID: 
PXD006847), sample annotations were infered from column names.

\strong{Peptide expression data: \code{zhu2018NC_hela_peptide}}

We processed the MaxQuant output \code{CulturedCells_peptides.txt} as 
follows:
\itemize{
  \item Extract the phenotype data (sample type, and number of cells when 
  single cells) from the intensity column names.
  \item Remove contaminants and reverse hits 
  \item Remove peptides with a posterior error probability < 0.01
  \item Replace 0's by NA's
}
We finally formated the data to an \code{\link{SingleCellExperiment}} 
object. The samples in this data set are either a cell lysate 
(\code{"lysate"}), HeLA cells (\code{"cells"}), blank samples 
(\code{"blank"}), or cancer cell lines (\code{"MCF7"} and \code{"THP1"}).
}
\references{
Zhu, Ying, Paul D. Piehowski, Rui Zhao, Jing Chen, Yufeng Shen, Ronald J. 
Moore, Anil K. Shukla, et al. 2018. “Nanodroplet Processing Platform for Deep 
and Quantitative Proteome Profiling of 10-100 Mammalian Cells.” Nature 
Communications 9 (1): 882. \href{http://dx.doi.org/10.1038/s41467-018-03367-w}{DOI}
}
\keyword{datasets}
