% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhu2018NC_islets_peptide}
\alias{zhu2018NC_islets_peptide}
\title{laser dissection + nanoPOTS: T1D islets (Zhu et al. 2018, Nat Comm)}
\format{\itemize{
  \item \code{zhu2018NC_islets_peptide}: a \code{\link{SingleCellExperiment}} 
  with peptide expression levels for 18611 peptides x 18 samples
}
See Details for information about data collection.}
\source{
The raw data, the idenfication and quantification results can be found in the
PRIDE repository 
\href{ftp://ftp.pride.ebi.ac.uk/pride/data/archive/2018/01/PXD006847}{PXD006847}
}
\usage{
data("zhu2018NC_islets_peptide")
}
\description{
Near single-cell proteomics data produced and published by Zhu et al. 
(see references). It contains quantitative information for 18 human pancreas 
samples. The samples are pancreatic islets laser dissected from pancreatic 
tissue slices from pancreata recovered from organ donors through the 
JDRFNetwork for Pancreatic Organ Donors with Diabetes (nPOD) program. 9 
samples out 18 are control islets, the 9 others are from type 1 diabetes 
(T1D) patients.
}
\details{
Peptide data were collected from the PRIDE repository (accession ID: 
PXD006847), sample annotations were infered from column names.

\strong{Peptide expression data: \code{zhu2018NC_islets_peptide}}

We processed the MaxQuant output \code{Islet_t1d_ct_peptides.txt} as 
follows:
\itemize{
  \item Extract the phenotype data (sample type) from the intensity column 
  names.
  \item Remove contaminants and reverse hits 
  \item Remove peptides with a posterior error probability < 0.01
  \item Replace 0's by NA's
}
We finally formated the data to an \code{\link{SingleCellExperiment}} 
object. The samples in this data set are either a cell lysate 
(\code{"lysate"}), HeLA cells (\code{"cells"}), blank samples 
(\code{"blank"}), or cancer cell lines (\code{"MCF7"} and \code{"THP1"}).
}
\references{
Zhu, Ying, Paul D. Piehowski, Rui Zhao, Jing Chen, Yufeng Shen, Ronald J. 
Moore, Anil K. Shukla, et al. 2018. “Nanodroplet Processing Platform for Deep 
and Quantitative Proteome Profiling of 10-100 Mammalian Cells.” Nature 
Communications 9 (1): 882. \href{http://dx.doi.org/10.1038/s41467-018-03367-w}{DOI}
}
\keyword{datasets}
