% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dou2019_boosting}
\alias{dou2019_boosting}
\alias{dou2019_boosting_protein}
\title{FACS + nanoPOTS + TMT multiplexing: testing boosting ratios (Dou et al. 2019)}
\format{\code{dou2019_boosting} contains 1 set of data:
\itemize{
  \item \code{dou2019_boosting_protein}: a SingleCellExperiment with protein 
  expression levels for 1436 proteins x 60 cells.
}
See Details for information about data collection.}
\source{
The peptide and raw data can be downloaded from the massIVE repository 
\href{ftp://massive.ucsd.edu/MSV000084110/}{MSV000084110}.

The protein data can be downloaded from the 
\href{https://pubs.acs.org/doi/10.1021/acs.analchem.9b03349}{ACS Publications}
website (Supplementary information section).
}
\usage{
dou2019_boosting 
data("dou2019_boosting_protein")
}
\description{
Single-cell proteomics using nanoPOTS combined with TMT isobaric labeling. 
The cell types are either "Raw" (macrophage cells), "C10" 
(epihelial cells), or "SVEC" (endothelial cells). Each cell is replicated 2 
or 3x. Each cell type was run using 3 levels of boosting: 0 ng (no boosting), 
5 ng or 50 ng. When boosting was applied, 1 reference well and 1 boosting 
well were added, otherwise 1 empty well was added. Each boosting setting 
(no boosting, 5ng, 50ng) was run twice.
}
\details{
The peptide data were downloaded from the massIVE database (accession ID: 
MSV000084110) and the protein data were downloaded from the supplementary 
information section of the publisher's website (see sources). 

\strong{Peptide expression data: \code{dou2019_boosting_peptide}}

TODO

\strong{Protein expression data: \code{dou2019_boosting_protein}}

The data set is published as the supplementary data set 2 by Dou et al. in 
2019 (see references). The expression matrix contains 1436 proteins x 60 
single cells. The spreadsheet is called \code{ac9b03349_si_004.xlsx} and contains 7 sheets 
from which we took the 2nd, 4th and 6th sheets (named \code{"01 - No Boost 
raw data"}, \code{"03 - 5ng boost raw data"}, \code{"05 - 50ng boost raw 
data"}, respectively). It corresponds to the assembly of MSGF+ 
identifications and MASIC reporter assemblies. The data is then isotope 
corrected and sum rolled-up to the protein level. Contaminant and reverse hit 
are removed from this table. This is all performed by the authors. We matched 
the proteins between the three boosting settings and combined all data in a 
single table. The boosting quantities are kept as phenotype data. The data 
set is formated to an \code{\link{SingleCellExperiment}} object.
}
\references{
Dou, Maowei, Geremy Clair, Chia-Feng Tsai, Kerui Xu, William B. 
Chrisler, Ryan L. Sontag, Rui Zhao, et al. 2019. “High-Throughput Single Cell 
Proteomics Enabled by Multiplex Isobaric Labeling in a Nanodroplet Sample 
Preparation Platform.” Analytical Chemistry, September 
(\href{https://doi.org/10.1021/acs.analchem.9b03349}{DOI}).
}
\keyword{datasets}
