% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{scpdata}
\alias{scpdata}
\title{List the available data sets in the scpdata package}
\usage{
scpdata()
}
\value{
An object of class "packageIQR" containing all available data sets.
}
\description{
The function lists the available datasets in the package.
}
\details{
See the documentation of a particular data set for more information (eg 
\code{?specht2019}).
}
\examples{
scpdata()

}
