% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{specht2019}
\alias{specht2019}
\alias{specht2019_protein}
\alias{specht2019_peptide}
\alias{specht2019_peptide2}
\title{FACS + SCoPE2: comparing macrophages against monocytes (Specht et al. 2019)}
\format{\code{specht2019} contains 2 sets of data:
\itemize{
  \item \code{specht2019_peptide}: an MSnSet with peptide expression levels 
  for 6787 peptides x 356 cells.
  \item \code{specht2019_peptide2}: an MSnSet with peptide expression levels 
  for 13543 peptides x 2420 samples. See details for more information about
  the sample phenotypes.
  \item \code{specht2019_protein}: an MSnSet with protein expression levels
  for 2316 proteins x 356 cells.
}
See Details for information about data collection.}
\source{
The original data can be downloaded from the 
\href{https://scope2.slavovlab.net/docs/data}{Slavov Lab} website.
}
\usage{
specht2019
data("specht2019_protein")
data("specht2019_peptide")
data("specht2019_peptide2")
}
\description{
Single cell proteomics data produced and published by Specht et al. from the 
Slavov Lab (see references). It contains quantitative information 356 cells. 
Cells can be either macrophages (n = 259) or monocytes (n = 97).
}
\details{
Three data files are provided by Specht and colleagues are:
\itemize{
  \item \code{Peptides-raw.csv}: Peptides x single cells at 1% FDR. The 
  first 2 columns list the corresponding protein identifiers and peptide 
  sequences and each subsequent column corresponds to a single cell. Peptide 
  identification is based on spectra analyzed by MaxQuant and is enhanced by 
  using DART-ID to incorporate retention time information. 
  \item \code{Proteins-processed.csv}: Proteins x single cells at 1% FDR, 
  imputed and batch corrected.
  \item \code{Cells.csv}: Annotation x single cells. Each column corresponds 
  to a single cell and the rows include relevant metadata, such as, cell type 
  if known, measurements from the isolation of the cell, and derivative 
  quantities, i.e., rRI, CVs, reliability.
 }
 
 Furthermore, in a personal discussion Harrison Specht shared the files:
 
 \itemize{
   \item \code{ev_updated.txt}: the MaxQuant output file of MS data 
   identification and quantification
   \item \code{annotation_fp60-97.csv}: sample phenotypic annotations
   \item \code{batch_fp60-97.csv}: batch annotations
   \item \code{te269088_lot_correction.csv}: TMT isotopic cross-contamination 
   correction matrix
 }
 
 \strong{Peptide expression data: \code{specht2019_peptide}}
 
 The \code{Peptides-raw.csv} data have already been partially processed by 
 the authors. The MS files were analyzed with MaxQuant and DART-ID and the 
 output \code{evidence.txt} file was parsed into R. This output was further 
 processed by the authors as follows:
 \itemize{
   \item Keep only the single cell runs (experiments FP94 and FP97)
   \item Remove out reverse hits and contaminants (identified by MaxQuant), 
   and contaminated spectra (\code{PIF > 0.8})
   \item Remove peptides with low identification score (\code{FDR >= 0.01} or 
   \code{PEP >= 0.02})
   \item Remove cells with less than 300 peptides
   \item Remove peptides that are more than 10\% the intensity of the carrier
   \item Divide peptide intensities in every channel by the reference channel
   \item Convert data to matrix format
   \item Zero or infinite intensities are replaced by \code{NA}'s
   \item Remove cells having \code{median CV > 0.43}, havuing 30th quantile of 
   the log10 transformed relative RIs smaller than -2.5, or having median of 
   the log10 transformed relative RI larger than -1.3
   \item Divide column (cells) with median intensity and divide rows with 
   mean intensity
   \item Remove rows (peptides) then columns (cells) that contain more than 
   99\% of missing data
   \item Log2 transform the data
 }
 
 The peptide data (\code{Peptides-raw.csv}) and the meta data 
 (\code{Cells.csv}) were combined into an \code{\link{MSnSet}} object.
 
 \strong{Peptide expression data: \code{specht2019_peptide2}}
 
 Inspired by the above steps, we processed the MaxQuant output
 \code{ev_updated.txt} as follows:
 \itemize{
   \item Keep only column of interest, that is the peptide and protein 
   information, the quantification information (TMT intensities), the 
   identification information (PEP, q-values, PIF,...) and the contamination
   information
   \item Remove contaminants and reverse hits
   \item Remove peptides with low identification metrics
   \item Remove runs with less than 300 identified peptides. 13 runs out of 
   245 were removed and were mainly blank or QC samples. 
   \item Some peptides were identified twice within the same experiment set 
   because of different charge states. The peptide identification with lowest 
   PEP is kept and the other(s) discarded.
   \item Intensity data is formated to a feature (peptide) x sample 
   (combination of experiment run and TMT channel) matrix.
   \item Peptide information is gathered in a feature data frame, only 
   peptide information common to all runs is kept (eg sequence, mass, 
   protein,...).
   \item Sample information is gathered in a phenotype data frame. The sample 
   information is extracted from the \code{annotation_fp60-97.csv} and 
   \code{batch_fp60-97.csv} files.
 }
 
 We finally formated the data to an MSnSet object. The samples in this data 
 set are either single cells, 10 cells, 100 cells, 1000 cells or QC. 
 
 
 \strong{Protein expression data: \code{specht2019_protein}}
 
 On top of the steps described above, the peptide expression data were further
 processed by the authors to the protein data through the following steps: 
 \itemize{
   \item Aggregate the peptide to their corresponding protein using their 
   median value.
   \item Subtracting rows by the averages and subtract column by the column
   median
   \item Impute missing values using K-nearest neighbors (k = 3)
   \item Correct for batch effect using the \code{\link{ComBat}} algorithm.
 }
 This leads to the \code{Proteins-processed.csv} file. This protein 
 expression data are combined with the meta data (\code{Cells.csv}) into an 
 \code{\link{MSnSet}} object.#'
}
\references{
Specht, Harrison, Edward Emmott, Toni Koller, and Nikolai Slavov. 
2019. “High-Throughput Single-Cell Proteomics Quantifies the Emergence of 
Macrophage Heterogeneity.” bioRxiv (\href{https://doi.org/10.1101/665307}{DOI}).
}
\keyword{datasets}
